Added manual mock for walk
